Add Cultureinfo.invariantculture to input amount. change createAccoun…

+++ LifeCycleBank/Bank.cs
@@ -109,7 +109,7 @@ namespace LifeCycleBank
                 return "false";
         }
 
-        public string CreateAccount(ICustomer customerId, int balance)
+        public string CreateAccount(ICustomer customerId, decimal balance)
         {
             try
             {
